+++ js/main.js
@@ -1,6 +1,6 @@
 import {getData} from './api.js';
 import {showErrorMessage} from './util.js';
-import {initThumbnails} from './create-photos.js';
+import {initThumbnails} from './filter.js';
 import {addFormEventsValidation} from './form.js';
 
 getData()
